kgapp/requirements.txt
amqp==5.1.1
anyascii==0.3.1
arrow==1.2.3
async-timeout==4.0.2
Babel==2.11.0
beautifulsoup4==4.7.1
bibtexparser==1.1.0
billiard==3.6.4.0
binaryornot==0.4.4
blinker==1.5
celery==5.1.2
celery-once==3.0.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookiecutter==1.7.3
Cython==0.29.33
dnspython==2.2.1
email-validator==1.1.3
et-xmlfile==1.1.0
eventlet==0.33.0
filedepot==0.8.0
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-WTF==0.14.3
future==0.18.2
greenlet==2.0.1
html5lib==1.1
idna==3.4
ijson==2.4
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.2.0
kombu==5.2.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.0.1
nltk==3.6.5
nose==1.3.7
numpy==1.24.1
openpyxl==3.0.10
oxrdflib==0.3.1
pandas==1.5.2
passlib==1.7.4
pip==22.0.2
poyo==0.5.0
prompt-toolkit==3.0.36
psutil==5.9.4
puremagic==1.14
PyJWT==2.6.0
pyoxigraph==0.3.10
pyparsing==3.0.9
pyshp==2.3.1
python-dateutil==2.8.2
python-slugify==7.0.0
pytidylib==0.3.2
pytz==2022.7
rdflib==6.0.2
rdflib-jsonld==0.6.2
redis==4.4.2
redislite==6.2.805324
regex==2022.10.31
requests==2.28.2
requests-testadapter==0.3.0
sadi==0.5.7
scipy==1.10.0
sdd2rdf==1.0.4
setlr==0.2.18
setuptools==59.6.0
six==1.16.0
soupsieve==2.3.2.post1
speaklater==1.3
text-unidecode==1.3
toposort==1.9
tqdm==4.64.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
whyis==2.0.4
WTForms==3.0.1
xlrd==2.0.1

mod-wsgi==4.9.0
